docs: update readme template, examples, drop http
