+++ Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["pam", "pam-sober", "pam-http"]
+members = ["pam"]
