--- pam-http/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "pam-http"
-version = "0.1.0"
-authors = ["Anthony Nowell <anowell@gmail.com>"]
-
-[lib]
-name = "pam_http"
-crate-type = ["cdylib"]
-
-[dependencies]
-pam-bindings = { path = "../pam/" }
-reqwest = { version = "0.11.3", features = ["blocking"] }
