--- pam-http/Justfile
@@ -1,12 +1,0 @@
-
-all:
-    cargo build
-
-install:
-    @cargo build --release
-    sudo cp conf/http-auth /etc/pam.d/
-    sudo cp ../target/release/libpam_http.so /lib/security/pam_http.so
-
-test:
-    @just install
-    gcc -o ../target/pam_test test.c -lpam -lpam_misc
