--- pam-sober/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "pam-sober"
-version = "0.1.0"
-authors = ["Anthony Nowell <anowell@gmail.com>"]
-
-[lib]
-name = "pam_sober"
-crate-type = ["cdylib"]
-
-[dependencies]
-pam-bindings = { path = "../pam/" }
-rand = "0.8.4"
