--- pam-sober/Justfile
@@ -1,12 +1,0 @@
-
-all:
-    cargo build
-
-install:
-    @cargo build --release
-    sudo cp conf/sober-auth /etc/pam.d/
-    sudo cp ../target/release/libpam_sober.so /lib/security/pam_sober.so
-
-test:
-    @just install
-    gcc -o ../target/pam_test test.c -lpam -lpam_misc
