--- pam-sober/conf/sober-auth
@@ -1,2 +1,0 @@
-auth sufficient pam_sober.so
-account sufficient pam_sober.so
