+++ pam/src/lib.rs
@@ -1,28 +1,40 @@
-//! Interface to the pluggable authentication module framework (PAM).
+//! Simplified PAM module creation in Rust.
 //!
-//! The goal of this library is to provide a type-safe API that can be used to
-//! interact with PAM.  The library is incomplete - currently it supports
-//! a subset of functions for use in a pam authentication module.  A pam module
-//! is a shared library that is invoked to authenticate a user, or to perform
-//! other functions.
+//! ```rust
+//! use pam::constants::{PamFlag, PamResultCode};
+//! use pam::module::{PamHandle, PamHooks};
+//! use std::ffi::CStr;
 //!
-//! For general information on writing pam modules, see
-//! [The Linux-PAM Module Writers' Guide][module-guide]
+//! struct AliceOnly;
+//! pam::pam_hooks!(AliceOnly);
 //!
-//! [module-guide]: http://www.linux-pam.org/Linux-PAM-html/Linux-PAM_MWG.html
+//! impl PamHooks for AliceOnly {
+//!     fn sm_authenticate(pamh: &mut PamHandle, _args: Vec<&CStr>, _flags: PamFlag) -> PamResultCode {
+//!         let username = match pamh.get_user(None) {
+//!             Ok(username) => username,
+//!             Err(e) => {
+//!                 eprintln!("failed to get username, error code: {e:?}");
+//!                 return e;
+//!             }
+//!         };
 //!
-//! A typical authentication module will define an external function called
-//! `pam_sm_authenticate()`, which will use functions in this library to
-//! interrogate the program that requested authentication for more information,
-//! and to render a result.  For a working example that uses this library, see
-//! [toznyauth-pam][].
+//!         match username.as_str() {
+//!             "alice" => PamResultCode::PAM_SUCCESS,
+//!             _ => PamResultCode::PAM_AUTH_ERR,
+//!         }
+//!     }
+//! }
+//! # fn main() {}
+//! ```
 //!
-//! [toznyauth-pam]: https://github.com/tozny/toznyauth-pam
+//! # Introduction
 //!
-//! Note that constants that are normally read from pam header files are
-//! hard-coded in the `constants` module.  The values there are taken from
-//! a Linux system.  That means that it might take some work to get this library
-//! to work on other platforms.
+//! The Pluggable Authentication Modules (PAM) framework enables systems to
+//! authenticate users and perform other functions by composing PAM modules,
+//! which are distributed as shared libraries.
+//!
+//! The goal of this library is to provide a simple, type-safe API to write
+//! these modules.
 
 extern crate libc;
